Added the new functionality from main to the addtoteamviewmodel
