+++ app/src/main/java/com/example/pokedex/repositories/TeamsRepository.kt
@@ -8,6 +8,7 @@ import androidx.lifecycle.asLiveData
 import com.example.pokedex.data.DatabaseService
 import com.example.pokedex.dependencyContainer.DependencyContainer
 import com.example.pokedex.dataClasses.Pokemon
+import com.example.pokedex.dataClasses.PokemonTypeResources
 import com.example.pokedex.dataClasses.Team
 import com.google.gson.Gson
 import kotlinx.coroutines.CoroutineScope
@@ -22,6 +23,7 @@ import kotlinx.coroutines.launch
 private val Context.dataStore by preferencesDataStore(name = "team_preferences")
 
 class TeamsRepository(private val context: Context) {
+    private val recentlySearchedRepository = DependencyContainer.recentlySearchedRepository
     private val databaseService = DatabaseService("teams", Team::class.java)
     private val connectivityRepository = DependencyContainer.connectivityRepository
     private var hasInternet = connectivityRepository.isConnected.asLiveData()
@@ -157,6 +159,29 @@ class TeamsRepository(private val context: Context) {
         val preferences = context.dataStore.data.first()
         val teamsJson = preferences[TEAMS_KEY] ?: "[]"
         return gson.fromJson(teamsJson, Array<Team>::class.java).toList()
+    }
+
+    suspend fun fetchTeamSuggestions(teamName: String) {
+        val team = getTeam(teamName)
+
+        val teamTypes = mutableListOf<String>()
+        val allTypes = PokemonTypeResources().getAllTypes()
+        val missingTypes = mutableListOf<String>()
+
+        if (team != null) {
+            for (pokemon in team.pokemons) {
+                for (types in pokemon.types) {
+                    teamTypes.add(types.type.name)
+                }
+            }
+        }
+        for (type in allTypes) {
+            if (!teamTypes.contains(type)) {
+                missingTypes.add(type)
+            }
+        }
+
+        recentlySearchedRepository.searchPokemonByNameAndFilterWithSort("",0, missingTypes, "", -1)
     }
 
 }
